skeleton test script for server
